Rename PluginsArray to PluginArray
