+++ docs/src/content/packages/core/plugins.mdx
@@ -8,6 +8,6 @@ slug: core/plugins
 
 <CodeExample lang="ts">
   ```ts
-  import type { Plugin } from "../modules/PluginsArray";
+  import type { Plugin } from "../modules/PluginArray";
   ```
 </CodeExample>
