+++ packages/core/src/js/modules/index.ts
@@ -3,5 +3,5 @@ export * from "./EventEmitter";
 export * from "./FocusableArray";
 export * from "./FocusTrap";
 export * from "./localStore";
-export * from "./PluginsArray";
+export * from "./PluginArray";
 export * from "./StackArray";
