+++ packages/core/src/js/plugins/cssConfig.ts
@@ -1,5 +1,5 @@
 import { getCustomProps } from "../helpers";
-import type { Plugin } from "../modules/PluginsArray";
+import type { Plugin } from "../modules/PluginArray";
 import type { CollectionEntry } from "../CollectionEntry";
 
 export interface CSSConfig {
