+++ packages/core/src/js/plugins/debug.ts
@@ -1,5 +1,5 @@
 import { getValue } from "../utilities";
-import type { Plugin } from "../modules/PluginsArray";
+import type { Plugin } from "../modules/PluginArray";
 import type { Collection } from "../Collection";
 import type { CollectionEntry } from "../CollectionEntry";
