+++ packages/core/src/js/plugins/propStore.ts
@@ -1,6 +1,6 @@
 import { getValue } from "../utilities";
 import { localStore } from "../modules";
-import type { Plugin } from "../modules/PluginsArray";
+import type { Plugin } from "../modules/PluginArray";
 import type { CollectionEntry } from "../CollectionEntry";
 
 export interface PropStorePlugin extends Plugin {
